+++ CHANGELOG.md
@@ -5,6 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2018-12-26
+## [1.0.4] - 2018-01-09
+### Added
+- Add support for new options from documentation **v1.846** (2018-01-08)
+    - Add **rec_house_number** attribute
+    - Add **rec_block** attribute
+    - Add **rec_enterance** attribute
+    - Add **rec_floor** attribute
+    - Add **rec_flat_number** attribute  
+### Changed
+- Move repo to [**@inspirum**](https://github.com/inspirum) account
+
+
+## [1.0.3] - 2018-01-02
+### Fixed
+- Fix links url
+
+
+## [1.0.2] - 2018-01-02
+### Changed
+- Change composer package name to `inspirum/balikobot`
+
+
+## [1.0.1] - 2018-12-30
+### Fixed
+- Fixed tests
+
+
+## [1.0.0] - 2018-12-27
 ### Added
 - Full support for all requests and options in documentation **v1.845** (2018-11-29)